+++ packages/terra-i18n/src/translationsLoaders.js
@@ -34,4 +34,4 @@ const loadTranslations = (locale, callback, scope) => {
   }
 };
 
-module.exports = loadTranslations;
+export default loadTranslations;